+++ examples/torch/td3_hopper.py
@@ -14,7 +14,8 @@ from garage.torch.exploration_strategies.base import \
     PolicyWrappedWithExplorationStrategy
 from garage.torch.exploration_strategies.gaussian_strategy import GaussianStrategy
 from garage.torch.launchers.launcher_util import setup_logger
-from garage.torch.algos.networks import FlattenMlp, TanhMlpPolicy
+from garage.torch.algos.networks import FlattenMlp
+from garage.torch.policies import TanhMlpPolicy
 from garage.torch.algos.td3.td3 import TD3
