+++ garage/torch/policies/__init__.py
@@ -1,4 +1,8 @@
 from garage.torch.policies.policies import TanhGaussianPolicy
 from garage.torch.policies.policies import MakeDeterministic
+from garage.torch.policies.policies import MlpPolicy
+from garage.torch.policies.policies import TanhMlpPolicy
 
-__all__ = ["TanhGaussianPolicy", "MakeDeterministic"]
+__all__ = [
+    "TanhGaussianPolicy", "MakeDeterministic", "MlpPolicy", "TanhMlpPolicy"
+]
