+++ garage/torch/policies/policies.py
@@ -2,6 +2,7 @@ import numpy as np
 import torch
 from torch import nn as nn
 
+from garage.torch.algos.data_management.normalizer import TorchFixedNormalizer
 from garage.torch.policies.base import ExplorationPolicy, Policy
 from garage.torch.algos.distributions import TanhNormal
 from garage.torch.algos.networks import Mlp
@@ -134,3 +135,39 @@ class MakeDeterministic(Policy):
     def get_actions(self, observations):
         return self.stochastic_policy.get_actions(
             observations, deterministic=True)
+
+
+class MlpPolicy(Mlp, Policy):
+    """
+    A simpler interface for creating policies.
+    """
+
+    def __init__(self,
+                 *args,
+                 obs_normalizer: TorchFixedNormalizer = None,
+                 **kwargs):
+        self.save_init_params(locals())
+        super().__init__(*args, **kwargs)
+        self.obs_normalizer = obs_normalizer
+
+    def forward(self, obs, **kwargs):
+        if self.obs_normalizer:
+            obs = self.obs_normalizer.normalize(obs)
+        return super().forward(obs, **kwargs)
+
+    def get_action(self, obs_np):
+        actions = self.get_actions(obs_np[None])
+        return actions[0, :], {}
+
+    def get_actions(self, obs):
+        return self.eval_np(obs)
+
+
+class TanhMlpPolicy(MlpPolicy):
+    """
+    A helper class since most policies have a tanh output activation.
+    """
+
+    def __init__(self, *args, **kwargs):
+        self.save_init_params(locals())
+        super().__init__(*args, output_activation=torch.tanh, **kwargs)
